test: add docker run (#126)
